Split tests/API/XCCDF/unittests/all.sh to separate test cases
The `all.sh` script combines roughly 80 test cases in a single test case.
When running unit tests, it's difficult to debug fails.  It's also
difficult to orient in output. It isn't possible to easily run selected
single test case. If we split the test cases, it will be easier to work
with the tests. For example, you can run `ctest -R <test_name>`.  The
`all.sh` script can be considered an unnecessary abstraction layer.