Check local filesystems when using 'filepath' element
The recurse_file_system="local" doesn’t work with file_object and
textfilecontent54_object where <filepath> element is used to define the
file path and operation=”equals” is used on filepath.  As of OVAL
specification of FileBehaviors element:
https://oval.mitre.org/language/version5.10/ovaldefinition/documentation/unix-definitions-schema.html#FileBehaviors
"'recurse_file_system' defines the file system limitation of any
searching and applies to all operations as specified on the path or
filepath entity." Notice the words “any searching” and notice the words
“all operations”. We can solve this bug by performing a similar check
as we do in oval_fts_read_recurse_path.