Do not show misleading "pass" or "fail" in HTML report
The report showed "pass"/"fail" in section "OVAL details". But in OVAL
"pass"/"fail" doesn't exist. Instead, in OVAL we have
"true"/"false"/"unknown"/"error"/"not evaluated"/"not applicable". It
showed "pass" when the definition was "true" and "fail" elsewhere.
Also, it didn't show the results of the tests, but it showed the results
of the whole definition at each test - the test results was shown as
pass when the whole definition was true even if the specific test wasn't
evaluated as true.  Also fixes the misleading "because ..." prose,
instead shows a header based on the condition we use in XSLT template
(whether OVAL items were or were not collected for the given test).