Fix misguiding use of strncpy
Here we actually don't want to NUL-terminate the string and we
don't want to add zero bytes padding, because we are going to
replace it by a '\n' right away. `memcpy` is more accurate here because
it never adds any zero bytes From the previous code we are sure
that result array is large enough to fit the string.

Addressing:
/home/jcerny/openscap/src/common/err_queue.c: In function ‘err_queue_to_string’:
/home/jcerny/openscap/src/common/err_queue.c:133:4: warning: ‘strncpy’ output truncated before terminating nul copying as many bytes from a string as its length [-Wstringop-truncation]
    strncpy(pos, err->desc, desc_len);
    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/jcerny/openscap/src/common/err_queue.c:132:28: note: length computed here
    const size_t desc_len = strlen(err->desc);
                            ^~~~~~~~~~~~~~~~~