Drop bogus VERIFY_SIGNATURE and VERIFY_DIGEST checks
VERIFY_SIGNATURE and VERIFY_DIGEST are not independent verification
checks, these checks are performed internally by rpm and failure in
either will cause the entire header failing to load. These flags allow
disabling that verification, but this doesn't make sense for openscap
and doesn't work this way in rpm >= 4.15 anyway.