Prevent broken output in "--progress" mode
In "--progress" mode we print out the rule id and ': ' before the rule
evaluation starts and the rest of the line after the rule evaluation
finishes. If any warning occurs during the evaluation, it's printed in
the middle, which looks ugly. We can fix this by printing the
whole line at the same moment, after the rule evaluation finishes.
Also, the warnings should use dW macros in order to be able to turn them
on/off using `--verbose` option.
Related to https://bugzilla.redhat.com/show_bug.cgi?id=1640889