Fix reporting of multi-check rules in HTML report
Previously, only the first check of a multi-check rule was reported and
the rest was not displayed at all. The patch fixes both Rule Overview
and Result Details sections. It fixes also the OVAL details for these
rules. Only the details of the evaluated definition will be displayed
in each rule details table. Also, the OVAL definition ID and title
will be shown there.