Improve printed HTML report
There was a problem that the OVAL details section got cropped if it was
longer than 1 page of paper. That happened because the whole OVAL
details section was inside a table row (inside <tr> element) and table
rows are not allowed by Bootstrap to have a page break inside. Also, it
had set display:inline-block which also seems to prevent line breaks.
It can be solved by moving the OVAL details section outside the big
table and then remove the display:inline-block property from CSS class.

The patch also improves padding around the rule details panel body to
prevent melding the border lines together, which was very ugly with
failed rules where the gray border mixed with the red border.