Do not drop empty lines from Ansible remediations
We are removing empty lines from generated Ansible Playbook.  It looks
nice, but there is a catch.

Ansible tasks are a YAML snippets. So they contain YAML block scalars,
indicated by a pipe (|). Within these block scalars, the newlines should
be preserved.  By removing the empty lines from the scalars we are
actually changing the meaning of the scalars.

Unfortunately, OpenSCAP doesn't understand YAML, but only reads the
Ansible fix line-by-line as s string. That means it doesn't know when it
reads a block scalar. Therefore we need to stop removing the empty lines
anywhere.

This commit also adds a small test.

Resolves: RHBZ#1795563