Fix sed expression
The purpose of this test is to test if probe doesn't collect data
from the host. The comment in oval:0:def:2 in
test_probes_yamlfilecontent_offline_mode.xml says: "make sure file in
host doesn't match" and the expected result is false. The corresponding
state expects value of value_of to be `outstance`. Based on this,
the sed command should modify the file in host, not in guest, because
if the file in guest is modified the object and state will match and
definition result will be true.