Remove useless warnings
These warnings give zero value to the user, because they don't specify
what went wrong. Moreover, the severity of the message depends on the
parameters of the macro. For example, in environemntvariable58 probe the
macro is used in a way that "Entity has no value" is a valid and
expected situation, because if the pid element is set to xsi:nil the PID
should be set to the PID of the calling process.

Resolves: RHBZ#1764139