Query also unit-type-specific interfaces
All Unit objects implement the generic org.freedesktop.systemd1.Unit
interface. But, depending on the unit type they also implement one
unit-type-specific interface. The list of unit types and interfaces
can be found at:
https://www.freedesktop.org/wiki/Software/systemd/dbus/
We will decide which interface to query based on the second part
of the unit name (after the dot).