Fixes "passing an object that undergoes default argument promotion to 'va_start' has undefined behavior [-Wvarargs]"
The parameter parmN is the identifier of the rightmost parameter in
the variable parameter list in the function definition (the one just
before the ...). If the parameter parmN is declared with the register
storage class, with a function or array type, or with a type that
is not compatible with the type that results after application of
the default argument promotions, the behavior is undefined.