Call xccdf_session_create_arf_source only once
To understand this change it is important to know how the function
`xccdf_session_create_arf_source` works. It creates the result data
stream oscap_source stucture, but it also adds it to the xccdf_session.
If there is already the result data stream in the xccdf_session, it
doesn't create a new one, but it just returns a pointer to the existing
one. That means multiple subsequent calls for
xccdf_session_create_arf_source will return the same pointer. Therefore
we don't have to call it twice, the second call will give us the same
oscap_source.