Do not call if previous call failed
Do not call `rpmverify_collect_package_files_or_directories` a second
time with RPMTAG_DIRNAMES if the call with RPMTAG_BASENAMES resulted in
error (ret = -1)?  In the original code with the for loop, whenever
there was an error it would go to cleanup, exiting the loop and stop
collecting package files or directories.