Change local and remote file system detection
The previous way was unreliable because some remote file systems were
detected as local. For example, the SMB mounts that used paths with
backslashes were not recognized as remote file systems.  The `getmntent`
function returns a file system type, so we can use it for detection of
remote file systems.

The new solution is inspired by systemd, function `fstype_is_network`:
https://github.com/systemd/systemd/blob/21fd6bc263f49b57867d90d2e1f9f255e5509134/src/basic/mountpoint-util.c#L290

It also handles FUSE file systems. Also, we are canceling the TODO idea
of providing the data at the build time and removing the dead code (if
0).

Resolves: RHBZ#1869195