Fix -Wmaybe-uninitialized warning in xccdf_session.c
The variable tailoring_doc_timestamp should be initialized before the
function can jump to the cleanup label which attempts to free it. Fix
this by initializing it to NULL at the top of the function.

Also check the return value of malloc() when allocating space for
tailoring_doc_timestamp and handle the allocation not succeeding.