Test oscap under memory limit
This test runs evaluation of a profile under memory constrainsts. The
profile consists of 10 memory intensive rules. The memory limit is set
to 700 MB using cgroups. This test isn't a part of upstream test suite,
because it takes a longer time to complete.  Instead, it will be run by
a new nightly Jenkins job.