Add severity and role attributes
Adds severity and role attributes to rule-result elements in XCCDF
results. These attributes are optional in XCCDF 1.2 specification but if
the XCCDF is used in ARF the attributes are mandatory.  See SCAP 1.3
specification Section 4.5 item 10. We didn't create these attributes if
the input XCCDF didn't contain them. We can fix it by assuming the
default values. According to XCCDF 1.2 specification, the default value
of role element is "full" and the default value of severity element is
"unknown". The presence of role, weight and severity attributes is
checked by SCAPVAL. We had weight attribute but missing role and
severity caused error RES-392.

Fixes: #1630