Silence an unhandled enumeration value warning
The warning for oval_debian_evr_string_cmp() is due to the switch
statement not having cases for all possible values of oval_operation_t.
Fix this by adding a default case and calling oscap_seterr() there.
This does not fix a real bug, as the function already sets an error for
unhandled values immediately after leaving the switch statement.