Improve documentation about --stig-viewer option
The documentation of --stig-viewer option wasn't sufficient and
confused multiple users, for example RHBZ#1896957.

This change tries to describe the use case in which --stig-viewer
option is used in a greater detail.

The commit removes the information about --stig-viewer not working
with DISA STIG Viewer newer than 2.6 because it works fine with
version 2.11. The source of this information is unclear, probably
a confusion due to difference between using DISA benchmark and SSG.

Fixes: RHBZ#1918759