Ignore namespace in XPath expressions
The OVAL xmlfilecontent_test doesn't provide any way to register
namespaces or namespace prefixes. The only way to match namespaced XML
elements in an XPath expression would be to use wildcard and
local-name() function. However, the SCAP 1.3 validation test suite
expects in R2850 that the namespaces of the XML document are ignored.
And the existing SCAP content counts with this behavior.

Read this mailing list thread:
https://lists.cisecurity.org/empathy/thread/ZEJKSY7QMLVXNZ7QRQ555N7W2T2SDNTO

We haven't found any way to enable the behavior of ignoring namespaces
in libxml2. Therefore our solution is to remove namespaces from the
given XML file before examination. This is implemented using an one-off
XSLT transformation.

Fixes: #1638