tests: Ignore "Bad address" errors from the sysctl probe
On some systems there may be entries containing binary data that
won't fit into the buffer (8192). There is not much sense in this
data for OVAL processing so we in general are not interested in it.

For example (Ubuntu 18.04 on Arm64 / Travis):
E: oscap:     An error ocured when reading from
"/proc/sys/kernel/spl/hostid" (fp=0xaaaaea9d6cf0): l=0, 14, Bad address