Change reporting of SDS schematron validation
The schematron validation of an SCAP source data stream consists of
applying the source data stream schematron and then applying a component
type specific schematron for each component. The final result is a
combination of the results of all executed schematrons. Up until now,
only the results of respective schematrons were reported on stdout, but
the final result was represented only by the return code. Also, the
message for application of source data stream schematron was confusing
because it gave impression that it means the final result.