Skip schematron validation of some SCAP document types
For some SCAP document types eg. OCIL no schematron is available
therefore it isn't an error when it's not found but rather we
should skip it. This is important in the context of source data
stream component schematron validation because the OCIL component
present in the data stream would render the whole data stream invalid
because no schematron for OCIL exists.