OVAL/probes: Make 'proc' and 'sysfs' filesystems non-local
More recently, we've come to use the -l switch in the df
command to identify the "local" filesystems. We feel this
is probably a better reflection of the intention behind
the local attribute value.

The df utility considers 'proc' and 'sysfs' as "dummy" and
does not produce results for them.