Set the default value of role attribute
By default the value of role member of struct xccdf_refine_rule was set
to zero when created which happened usually when parsing a
xccdf:refine-rule element. But the value of 0 doesn't map to the default
value of the role attribute of xccdf:refine-rule element. The default
value of this attribute is "full" according to the XCCDF specification.
As a result when creating XCDF results or ARF results the role attribute
of xccdf:rule-result hasn't been created. That's a problem according to
scapval.
Fixes: #1729