Use the real benchmark title
Instead of the generic "OpenSCAP Security Guide" we will use the actual
Benchmark title. The title "OpenSCAP Security Guide" is confusing
because the guide is only generated using OpenSCAP but the source data
don't come from the OpenSCAP project. People are confused about the
source of the data, for example they report bugs to OpenSCAP when the
problem is in the SCAP content. Using the actual title taken from
the <xccdf:Benchmark> element will provide better flexibility.