Allow providing pre-downloaded components
OpenSCAP can download remote SCAP source data stream components from the
internet if the --fetch-remote-resources option is set. When this
command line option isn't set, it is possible to download the remote
component separately, save it as file to the filesystem which name is
equal to the name of the remote component and OpenSCAP would use this
local file instead of just skipping the checks. It currently works by
accident, it wasn't intended to work this way or at least I haven't
heard it before, but somobody found out that it works and documented it
as a KSC: https://access.redhat.com/solutions/5185891. However, it
produces warning messages if the approach used in the KSC is used.

This patch promotes the workaround to a feature. And it changes the
logic so that it won't print the warnings.

The main difference is that it won't work automatically, but only
on user's demand by providing --use-local-file which is the new
command line option introduced by this commit.

Fixes: rhbz#1970527