Another test for filehash58 probe
It renames the existing test case and also creates a new test case which
is almost identical with the existing test case but it uses different
hashing algorithms in the filehash58_test.  The reason for creating it
is that the existing test case uses the SHA-1 and MD5 alorithms. But we
introduced CMake options OPENSCAP_ENABLE_SHA1 and OPENSCAP_ENABLE_MD5
which allow to not build support for these algorithms. If OpenSCAP will
be built with these options the existing tests would be disabled and we
wouldn't test the filehash58 probe. We can expect that some downstreams
will turn the options OFF so they need an alternative test instead.