Provide a directory path
This will allow users to pass a directory path where downloaded files
are stored.

Currently, when it attempts to open a file, the path is resolved
relatively to the data stream file. For example, that means if the data
stream file is located in /usr/share/xml/scap you need to copy the file
there and you can have permission problems. Instead, we will pass the
directory and store the downloaded file to the directory.

The interface and command line options are updated to accept the
directory path. Moreover, we have realized that there can be multiple
remote components in a single source data stream. Therefore, we have
found a more suitable name for the option which would more fit that the
option accepts an argument and there can be multiple files.