Show a warning
When --local-files is used but the file which is expected to be
loaded instead of the remote data stream component doesn't exist
we will print a warning so that the user is informed and he would
know what file is expected.

Also added a simple test to test the aforementioned warning scenario.