Fix compilation of password probe on FreeBSD
After commit 062d027e6592 the read_password() and
_process_struct_passwd() functions use _PATH_LASTLOG and fgetpwent()
which are not portable to FreeBSD.

Fix this by having FreeBSD use completely separate functions to process
passwords as the differences are likely to grow over time.  Offline
processing on FreeBSD is not advertised as FreeBSD does not allow
setting a non-default password database.