Replace cuserid() with getpwuid(getuid())
The function cuserid() is not readily available on all platforms,
namely it requires linking in compatibility libraries on FreeBSD and it
does not seem to be available at all on macOS. To improve portability,
use getpwuid(getuid()) and return an empty string in the event that it
fails. This should still solve the problem mentioned in commit
b31cff1bc3a2, which is what led to cuserid() being added in the first
place.

From the bugs section of the GNU man page for cuserid(3):

"Nobody knows precisely what cuserid() does; avoid it in portable
programs. Or avoid it altogether: use getpwuid(geteuid()) instead,
if that is what you meant. Do not use cuserid()."