Allow empty /proc in offline mode
When scanning offline file systems the /proc might be empty. Currently,
OpenSCAP thinks that it means a permissions problems, which is often
true if it happens on a real system, but in offline mode it can be a
normal situation. We will not consider empty /proc an error in offline
mode.

The commit also includes a simple test case.

Inspired by eda9881e08f0398d1481f133fbb56c0080cfe9f3

Resolves: RHBZ #2008922