Patch the "bz2_datastream" test for env with /tmp on a dedicated partition or LV
The aim of this test is to check that the binary is capable of testing a
xccdf file. If the test fails (expected in the Docker environment
provided) or pass (expected if executed on a machine with /tmp
configured on a dedicated partition or logical volume), the current test
is succeeded.

The binary is working if the return code is 0 or 2.