Extend tests of the XCCDF requires and conflicts
The purpose of these new test cases is to preserve the current behavior
of rules with requires and conflicts elements when `oscap` is invoked
with the `--rule` option selecting specific rule or rules. The tests are
based on current behavior, not on any specification, as `--rule` option
is our custom feature, but at the same time we want notice behavior
changes in future.