Ignore selections and conflicts
If the user explicitly requests a rule using --rule, they expect all
the rules would be evaluated regardless selections and conflicts.
The use-case for the `--rule` option is development and testing
and in these situation it anyways behaves as an override.

However, this patch doesn't solve the "requires" element handling.
Required rules are still not added.