Issue a warning about required rules
If at least one `--rule` option is provided by the user, only the rules
listed as arguments of `--rule` options are evaluated. It ignores the
`requires` element in the rule. The required rules aren't added
automatically and aren't evaluated unless they're explicitely listed.
This way we keep the behavior of `--rule` option in previous versions of
OpenSCAP, because SSGTS relies on the fact that single occurence of
`--rule` options means that at most 1 rule is evaluated.

To avoid confusion, OpenSCAP will now report a warning that the rule
they are evaluating a requires another rule.