OVAL: probe: linux: fwupdsecattr_probe: OVAL testing for system firmware
Recently years, more and more firmware-level attack schemes are developed.
The malware can steal information or damage the system through lower-level
system firmware. For example, the malware can change the system firmware if
the secure boot is not properly set and the users are not aware of this until
significant data loss or system damage happens. If a kind of testing tool can
be used to test and system, the user will understand how dangerous is their
system and then update their system to prevent loss or damage. 

fwupd provides the firmware details, OVAL can be used to describe the test, and
scap generates the test report for the user. Also, the OVAL description links
the test result and CVE to notice the user about the importance of the
firmware vulnerability. Moreover, it could be a standard for hardware purchasing
and hardware configuration.

The proposed work can get the firmware and hardware configuration details from
fwupd through dbus. Based on the pre-generated OVAL description from fwupd
release, this work test the hardware and firmware configuration and generates
the reports for the users. The report is based on openSCAP format so the user
who is familiar with this can easily understand the report.

Reference
fwupd: https://github.com/fwupd/fwupd

Signed-off-by: Kate Hsuan <hpa@redhat.com>