Process CPE AL platforms if CPE dict isn't part of DS
With this change, we will be able to process SCAP source data
streams which use CPE AL platforms in XCCDF rules but at the same
time there is no CPE dictionary present in the SCAP source
data stream. Definition of a CPE dictionary isn't mandatory
for platforms to be evaluated.

A small test for this situation is introduced as well.

Fixes: #1962