Allow preferring SCE
This commit introduces a new environment variable
`OSCAP_PREFERRED_ENGINE`. This variable allows users to set a preffered
check engine for XCCDF rules. If a rule has multiple checks, the checks
for the preffered check engine will be used. Allowed values: `SCE`,
`OVAL`. If this variable is set to `SCE` and a rule has both SCE and
OVAL checks the SCE check will be used. If this variable is set to
`OVAL` and a rule has both SCE and OVAL checks the OVAL check will be
used. If this environment variable isn't set, the standard XCCDF
mechanism will be used for check selection.

This will allow us to explicitely prefer SCE checks when executing
`oscap` in Containerfile in the `podman build` process when building
hardened bootc images.

Also a small test is added to test this feature.