Update oscap-bootc to verify it runs in bootable container env
The script is updated to first verify that it runs in a bootable
container environment - `bootc` package must be installed and
`/run/.containerenv` file must exist which indicates we are running
inside a container. If it is not running inside a bootable container
environment it informs user and exits.

Another change is that installation of `openscap-engine-sce`
package has been moved from specfile into the script as the
script already installs other requirements which are needed
by SCE checks from CaC/content.